fix: validate org unit and org unit mode params [DHIS2-1585] [2.38]
