+++ dhis-2/dhis-web-api/src/main/java/org/hisp/dhis/webapi/controller/event/mapper/RequestToSearchParamsMapper.java
@@ -31,6 +31,7 @@ import static org.apache.commons.lang3.BooleanUtils.toBooleanDefaultIfNull;
 import static org.hisp.dhis.webapi.controller.event.mapper.EventParamsMapperUtils.getOrgUnitMode;
 import static org.hisp.dhis.webapi.controller.event.mapper.EventParamsMapperUtils.validateAccessibleOrgUnits;
 
+import com.google.common.base.Strings;
 import java.util.Collections;
 import java.util.Date;
 import java.util.HashMap;
@@ -71,6 +72,7 @@ import org.hisp.dhis.program.ProgramStatus;
 import org.hisp.dhis.query.QueryUtils;
 import org.hisp.dhis.schema.Schema;
 import org.hisp.dhis.schema.SchemaService;
+import org.hisp.dhis.security.Authorities;
 import org.hisp.dhis.security.acl.AclService;
 import org.hisp.dhis.trackedentity.TrackedEntityInstance;
 import org.hisp.dhis.trackedentity.TrackedEntityInstanceService;
@@ -247,6 +249,10 @@ public class RequestToSearchParamsMapper {
 
     if (!StringUtils.isEmpty(orgUnit) && ou == null) {
       throw new IllegalQueryException("Org unit is specified but does not exist: " + orgUnit);
+    }
+
+    if (orgUnitSelectionMode != null) {
+      validateOrgUnitMode(orgUnitSelectionMode, orgUnit, user);
     }
 
     OrganisationUnitSelectionMode orgUnitMode = getOrgUnitMode(ou, orgUnitSelectionMode);
@@ -479,5 +485,54 @@ public class RequestToSearchParamsMapper {
 
   public EventSearchParams map(TrackerEventCriteria eventCriteria) {
     return map(TRACKER_EVENT_CRITERIA_MAPPER.toEventCriteria(eventCriteria));
+  }
+
+  private void validateOrgUnitMode(
+      OrganisationUnitSelectionMode selectedOuMode, String orgUnit, User user) {
+
+    String violation = null;
+
+    switch (selectedOuMode) {
+      case ALL:
+        violation =
+            userCanSearchOuModeALL(user)
+                ? null
+                : "Current user is not authorized to query across all organisation units";
+        break;
+      case ACCESSIBLE:
+      case CAPTURE:
+        if (user == null) {
+          violation = "User is required for ouMode: " + selectedOuMode;
+        }
+        if (orgUnit != null) {
+          violation =
+              String.format(
+                  "ouMode %s cannot be used with orgUnits. Please remove the orgUnit parameter and try again.",
+                  selectedOuMode);
+        }
+        break;
+      case CHILDREN:
+      case SELECTED:
+      case DESCENDANTS:
+        violation =
+            orgUnit == null ? "Organisation unit is required for ouMode: " + selectedOuMode : null;
+        break;
+      default:
+        violation = "Invalid ouMode:  " + selectedOuMode;
+        break;
+    }
+
+    if (!Strings.isNullOrEmpty(violation)) {
+      throw new IllegalQueryException(violation);
+    }
+  }
+
+  private boolean userCanSearchOuModeALL(User user) {
+    if (user == null) {
+      return false;
+    }
+
+    return user.isSuper()
+        || user.isAuthorized(Authorities.F_TRACKED_ENTITY_INSTANCE_SEARCH_IN_ALL_ORGUNITS.name());
   }
 }
