+++ dhis-2/dhis-web-api/src/test/java/org/hisp/dhis/webapi/controller/event/EventRequestToParamsMapperTest.java
@@ -40,7 +40,9 @@ import static org.hisp.dhis.common.OrganisationUnitSelectionMode.CHILDREN;
 import static org.hisp.dhis.common.OrganisationUnitSelectionMode.DESCENDANTS;
 import static org.hisp.dhis.common.OrganisationUnitSelectionMode.SELECTED;
 import static org.hisp.dhis.utils.Assertions.assertContainsOnly;
+import static org.hisp.dhis.utils.Assertions.assertStartsWith;
 import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.when;
 
@@ -67,7 +69,6 @@ import org.hisp.dhis.user.CurrentUserService;
 import org.hisp.dhis.user.User;
 import org.hisp.dhis.webapi.controller.event.mapper.RequestToSearchParamsMapper;
 import org.hisp.dhis.webapi.controller.event.webrequest.EventCriteria;
-import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -165,7 +166,7 @@ class EventRequestToParamsMapperTest {
             null,
             null,
             null,
-            "orgunituid",
+            null,
             ACCESSIBLE,
             "teiUid",
             null,
@@ -268,7 +269,7 @@ class EventRequestToParamsMapperTest {
     eventCriteria.setOuMode(DESCENDANTS);
 
     IllegalQueryException exception =
-        Assertions.assertThrows(
+        assertThrows(
             IllegalQueryException.class, () -> requestToSearchParamsMapper.map(eventCriteria));
     assertEquals(
         "User does not have access to orgUnit: " + orgUnit.getUid(), exception.getMessage());
@@ -293,7 +294,7 @@ class EventRequestToParamsMapperTest {
         .thenReturn(orgUnitDescendants);
 
     IllegalQueryException exception =
-        Assertions.assertThrows(
+        assertThrows(
             IllegalQueryException.class, () -> requestToSearchParamsMapper.map(eventCriteria));
     assertEquals(
         "User does not have access to orgUnit: " + orgUnit.getUid(), exception.getMessage());
@@ -364,7 +365,7 @@ class EventRequestToParamsMapperTest {
     eventCriteria.setOuMode(CHILDREN);
 
     IllegalQueryException exception =
-        Assertions.assertThrows(
+        assertThrows(
             IllegalQueryException.class, () -> requestToSearchParamsMapper.map(eventCriteria));
     assertEquals(
         "User does not have access to orgUnit: " + orgUnit.getUid(), exception.getMessage());
@@ -387,7 +388,7 @@ class EventRequestToParamsMapperTest {
     eventCriteria.setOuMode(CHILDREN);
 
     IllegalQueryException exception =
-        Assertions.assertThrows(
+        assertThrows(
             IllegalQueryException.class, () -> requestToSearchParamsMapper.map(eventCriteria));
     assertEquals(
         "User does not have access to orgUnit: " + orgUnit.getUid(), exception.getMessage());
@@ -406,7 +407,6 @@ class EventRequestToParamsMapperTest {
 
     EventCriteria eventCriteria = new EventCriteria();
     eventCriteria.setProgram(program.getUid());
-    eventCriteria.setOrgUnit(orgUnit.getUid());
     eventCriteria.setOuMode(CAPTURE);
 
     EventSearchParams searchParams = requestToSearchParamsMapper.map(eventCriteria);
@@ -427,7 +427,6 @@ class EventRequestToParamsMapperTest {
 
     EventCriteria eventCriteria = new EventCriteria();
     eventCriteria.setProgram(program.getUid());
-    eventCriteria.setOrgUnit(orgUnit.getUid());
     eventCriteria.setOuMode(ACCESSIBLE);
 
     EventSearchParams searchParams = requestToSearchParamsMapper.map(eventCriteria);
@@ -519,10 +518,34 @@ class EventRequestToParamsMapperTest {
     eventCriteria.setOrgUnit(orgUnit.getUid());
 
     IllegalQueryException exception =
-        Assertions.assertThrows(
+        assertThrows(
             IllegalQueryException.class, () -> requestToSearchParamsMapper.map(eventCriteria));
     assertEquals(
         "User does not have access to orgUnit: " + orgUnit.getUid(), exception.getMessage());
+  }
+
+  @Test
+  void shouldFailWhenOrgUnitSuppliedAndOrgUnitModeAccessible() {
+    EventCriteria eventCriteria = new EventCriteria();
+    eventCriteria.setOrgUnit(orgUnit.getUid());
+    eventCriteria.setOuMode(ACCESSIBLE);
+    Exception exception =
+        assertThrows(
+            IllegalQueryException.class, () -> requestToSearchParamsMapper.map(eventCriteria));
+
+    assertStartsWith("ouMode ACCESSIBLE cannot be used with orgUnits.", exception.getMessage());
+  }
+
+  @Test
+  void shouldFailWhenOrgUnitSuppliedAndOrgUnitModeCapture() {
+    EventCriteria eventCriteria = new EventCriteria();
+    eventCriteria.setOrgUnit(orgUnit.getUid());
+    eventCriteria.setOuMode(CAPTURE);
+    Exception exception =
+        assertThrows(
+            IllegalQueryException.class, () -> requestToSearchParamsMapper.map(eventCriteria));
+
+    assertStartsWith("ouMode CAPTURE cannot be used with orgUnits.", exception.getMessage());
   }
 
   private OrganisationUnit createOrgUnit(String name, String uid) {
